Refactor names and fix style issues
