+++ frontend/src/components/completed-tasks-chart/completed-tasks-chart.options.js
