+++ frontend/src/components/profile-display/profile-display.css
@@ -1,11 +1,12 @@
 .profile-display-container {
     background-color: var(--color-blue);
     display: flex;
-    padding: 1.5em;
+    padding: 1em;
 }
 
 .profile-display-image {
-    width: 48px;
+    width: 60px;
+    height: 60px;
     border-radius: 50%;
     margin-right: 1em;
 }
@@ -20,6 +21,7 @@
 
 .profile-display-name {
     font-size: 1rem;
+    cursor: pointer;
 }
 
 .profile-display-button {
@@ -31,4 +33,8 @@
     cursor: pointer;
     font-weight: var(--font-weight-light);
     margin-top: 0.3em;
+}
+
+.profile-display-name:hover, .profile-display-button:hover {
+    text-shadow: -2px 0 #2577c4, 0 2px #2577c4, 2px 0 #2577c4, 0 -2px #2577c4;
 }
