+++ frontend/src/pages/profile/profile-page.module.css
@@ -44,6 +44,7 @@ h2{
 
 .userInfo img{
     width: 10vw;
+    height: 10vw;
     margin: 0;
 }
